refactor: extract shared CLI utilities from command handlers
Extracts duplicated code from diff.ts, apply.ts, and kubeconfig.ts into
shared utility modules: project loading, k0s provider setup, OCI helpers,
template validation, and k8s error detection. Fixes loose object[] typing
with K8sObjectType, hardens is_not_found_error regex, and adds exit code
documentation to the diff command.
