+++ src/cli/commands/kubeconfig.ts
@@ -1,10 +1,14 @@
 import * as path from 'node:path';
 import { is_success, success } from '../../core/index.js';
 import { create_kubeconfig_manager } from '../../k8s/kubeconfig.js';
-import { find_cluster, find_project_root, load_project } from '../../loader/index.js';
-import type { NodeListType } from '../../nodes/index.js';
 
 import { define_command } from '../command.js';
+import {
+  build_node_list,
+  create_k0s_provider_instance,
+  resolve_k0s_provider_options,
+} from '../utils/k0s-provider.js';
+import { load_and_resolve_project } from '../utils/project.js';
 
 function sanitize_filename_part(value: string): string {
   return value.replace(/[^a-zA-Z0-9._-]/g, '_');
@@ -46,30 +50,16 @@ export const kubeconfig_command = define_command({
 
     console.log('\n━━━ Kustodian Kubeconfig ━━━\n');
 
-    // Load project
-    console.log('Loading project configuration...');
-    const root_result = await find_project_root(project_path);
-    if (!is_success(root_result)) {
-      console.error(`  ✗ ${root_result.error.message}`);
-      return root_result;
-    }
-
-    const project_root = root_result.value;
-    const project_result = await load_project(project_root);
+    const project_result = await load_and_resolve_project(project_path, cluster_filter);
     if (!is_success(project_result)) {
-      console.error(`  ✗ ${project_result.error.message}`);
       return project_result;
     }
 
-    const project = project_result.value;
-
-    // Find target cluster
-    const loaded_cluster = find_cluster(project.clusters, cluster_filter);
+    const loaded_cluster = project_result.value.target_clusters[0];
     if (!loaded_cluster) {
-      console.error(`  ✗ Cluster '${cluster_filter}' not found`);
       return {
         success: false as const,
-        error: { code: 'NOT_FOUND', message: `Cluster '${cluster_filter}' not found` },
+        error: { code: 'NOT_FOUND', message: 'No clusters found in project' },
       };
     }
 
@@ -77,34 +67,10 @@ export const kubeconfig_command = define_command({
     console.log(`  → Cluster: ${cluster_name}`);
     console.log(`  → Nodes: ${loaded_cluster.nodes.length}`);
 
-    // Build NodeListType
-    const node_list: NodeListType = {
-      cluster: cluster_name,
-      nodes: loaded_cluster.nodes,
-      ...(loaded_cluster.cluster.spec.node_defaults?.label_prefix && {
-        label_prefix: loaded_cluster.cluster.spec.node_defaults.label_prefix,
-      }),
-    } as NodeListType;
-
-    // Load k0s provider
-    const k0s_package = 'kustodian-k0s';
-    const { create_k0s_provider } = await import(k0s_package);
-
-    const k0s_plugin = loaded_cluster.cluster.spec.plugins?.find(
-      (p) => p.name === 'k0s' || p.name === '@kustodian/plugin-k0s',
-    );
-    const plugin_config = k0s_plugin?.config ?? {};
-
-    const provider_options: Record<string, unknown> = {};
-    if (plugin_config['k0s_version']) {
-      provider_options['k0s_version'] = plugin_config['k0s_version'];
-    }
-    if (plugin_config['default_ssh']) {
-      provider_options['default_ssh'] = plugin_config['default_ssh'];
-    }
-    provider_options['cluster_name'] = loaded_cluster.cluster.metadata.code ?? cluster_name;
-
-    const provider = create_k0s_provider(provider_options);
+    // Build NodeListType and provider
+    const node_list = build_node_list(loaded_cluster);
+    const provider_options = resolve_k0s_provider_options(loaded_cluster, { include_all: false });
+    const provider = await create_k0s_provider_instance(provider_options);
 
     // Validate
     console.log('\n  → Validating cluster configuration...');
@@ -130,7 +96,7 @@ export const kubeconfig_command = define_command({
       tmpdir(),
       `kustodian-kubeconfig-${sanitize_filename_part(cluster_name)}.yaml`,
     );
-    await writeFile(temp_kubeconfig, kubeconfig_result.value as string, 'utf-8');
+    await writeFile(temp_kubeconfig, kubeconfig_result.value, 'utf-8');
 
     // Merge into ~/.kube/config
     console.log('  → Merging into ~/.kube/config...');
